refactor: 킹을 잡았을 때 바로 게임 종료 상태로 바뀌도록 수정
